pyproject.toml
[project]
name = "sequoia"
version = "0.1.0"
description = "Genesis sequoia"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "ipykernel>=6.29.5",
    "manifold>=0.2.0",
    "manifold3d>=3.0.1",
    "matplotlib>=3.10.0",
    "numpy>=2.2.3",
    "opencv-python>=4.11.0.86",
    "pandas>=2.2.3",
    "pip>=25.0.1",
    "pyntcloud>=0.3.1",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "selenium>=4.29.0",
    "torchvision>=0.21.0",
    "trimesh>=4.6.4",
    "tqdm>=4.67.1",
    "vtk>=9.4.1",
    "webdriver-manager>=4.0.2",
    "seaborn>=0.13.2",
    "pyglet<2",
    "scikit-learn>=1.6.1",
    "ipywidgets>=8.1.5",
    "nbformat>=5.10.4",
    "ipyvtklink>=0.2.2",
    "trame>=3.8.1",
    "trame-vtk>=2.8.15",
    "trame-vuetify>=2.8.1",
    "pyvista>=0.44.1",
    "trame-jupyter-extension>=2.1.4",
    "scikit-image>=0.25.2",
    "rtree>=1.4.0",
    "pytest>=8.3.5",
    "ipytest>=0.14.2",
    "fast-simplification>=0.1.9",
    "meshio>=5.3.5",
    "jupyter-cache>=1.0.1",
    "lightning>=2.5.1",
    "hydra-core>=1.3.2",
    "rootutils>=1.0.7",
    "pymeshlab==2023.12.post2",
    "torch>=2.6.0",
    "scipy>=1.15.2",
    "tabulate>=0.9.0",
    "wandb>=0.19.9",
    "omegaconf>=2.3.0",
    "optuna>=4.3.0",
    "pymcubes>=0.1.6",
    "open3d>=0.19.0",
    "pyrender>=0.1.45",
    "pyhocon>=0.3.61",
    "dotmap>=1.3.30",
    "ipdb>=0.13.13",
    "numpy-stl>=3.2.0",
    "mesh-to-sdf>=0.0.15",
]
